--- e2e-tests/react-app/utils/index.ts
@@ -1,1 +1,0 @@
-export * from './connect';
